// src/types/GridImagesProps.ts
export interface GridImagesProps {
  gallery: {
    first: {
      mobile: string;
      tablet: string;
      desktop: string;
    },
    second: {
      mobile: string;
      tablet: string;
      desktop: string;
    },
    third: {
      mobile: string;
      tablet: string;
      desktop: string;
    }
  }
}